refine query/mutation wrappers

+++ convex/notes/notes.ts
@@ -1,7 +1,8 @@
 import { deleteBlocksByNote } from '../blocks/blocks'
-import { deleteItemSharesAndBookmarks } from '../sidebarItems/cascadeDelete'
 import { checkItemAccess, requireItemAccess } from '../sidebarItems/validation'
 import { PERMISSION_LEVEL } from '../shares/types'
+import { deleteSidebarItemShares } from '../shares/itemShares'
+import { deleteItemBookmarks } from '../bookmarks/bookmarks'
 import { enhanceNoteWithContent } from './helpers'
 import type { CampaignMutationCtx, CampaignQueryCtx } from '../functions'
 import type { NoteWithContent } from './types'
@@ -22,15 +23,11 @@ export async function deleteNote(
   noteId: Id<'notes'>,
 ): Promise<Id<'notes'>> {
   const rawNote = await ctx.db.get(noteId)
-  const note = await requireItemAccess(
-    ctx,
-    ctx.campaign._id,
-    rawNote,
-    PERMISSION_LEVEL.FULL_ACCESS,
-  )
+  await requireItemAccess(ctx, rawNote, PERMISSION_LEVEL.FULL_ACCESS)
 
-  await deleteBlocksByNote(ctx, noteId, note.campaignId)
-  await deleteItemSharesAndBookmarks(ctx, note.campaignId, noteId)
+  await deleteBlocksByNote(ctx, noteId)
+  await deleteSidebarItemShares(ctx, noteId)
+  await deleteItemBookmarks(ctx, noteId)
   await ctx.db.delete(noteId)
 
   return noteId
